requirements.txt
# ============================================================================
# CareCrew - Automotive Predictive Maintenance Platform
# Python Dependencies
# ============================================================================

# Core AI Framework (includes LiteLLM for Ollama support)
crewai[tools]>=0.203.0,<1.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Web Framework (Dashboard Backend)
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=2.3.7

# Data Validation & Processing
pydantic>=2.0.0

# HTTP Requests & API Integration
requests>=2.31.0

# Optional: Database (if needed in future)
# sqlalchemy>=2.0.0
# alembic>=1.13.0

# Optional: Advanced AI Features (uncomment if needed)
# langchain>=0.1.0
# chromadb>=0.4.0

# Development & Testing (install separately with: pip install -r requirements-dev.txt)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.1.0
